requirements.txt
streamlit
pandas
openai
PyMuPDF
pdfplumber
fpdf2
python-dotenv
elevenlabs
speechrecognition
pyaudio
fpdf